# apps/api/docs/swagger.yaml
basePath: /
definitions:
...
  pet.PetPayload:
    properties:
      age:
        maximum: 100
        minimum: 0
        type: integer
      breed:
        maxLength: 100
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
      owner:
        maxLength: 100
        type: string
      species:
        enum:
        - dog
        - cat
        - bird
        - rabbit
        - fish
        - other
        type: string
    required:
    - name
    - species
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API REST para gestión de mascotas y usuarios.
  title: My Pets API
  version: "1.0"
paths:
...
  /api/v1/pets:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 'data: []Pet, total: int'
          schema:
            additionalProperties: true
            type: object
        "401":
          description: authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: error message
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: List all pets
      tags:
      - pets
    post:
      consumes:
      - application/json
      parameters:
      - description: Pet data
        in: body
        name: pet
        required: true
        schema:
          $ref: '#/definitions/pet.PetPayload'
      produces:
      - application/json
      responses:
        "201":
          description: 'data: Pet'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: error message
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Create a new pet
      tags:
      - pets
  /api/v1/pets/{id}:
    delete:
      parameters:
      - description: Pet ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: pet deleted'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: invalid id
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: pet not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: error message
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Delete a pet
      tags:
      - pets
    get:
      parameters:
      - description: Pet ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'data: Pet'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: invalid id
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: pet not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: error message
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Get a pet by ID
      tags:
      - pets
    put:
      consumes:
      - application/json
      parameters:
      - description: Pet ID
        in: path
        name: id
        required: true
        type: string
      - description: Pet data
        in: body
        name: pet
        required: true
        schema:
          $ref: '#/definitions/pet.PetPayload'
      produces:
      - application/json
      responses:
        "200":
          description: 'data: Pet'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: invalid id or validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: pet not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: error message
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Update an existing pet
      tags:
      - pets
...
securityDefinitions:
  CookieAuth:
    description: Cookie HttpOnly de sesión. Obtenida al hacer login o refresh. Se
      envía automáticamente por el navegador.
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"
...